Put our Banker class inside of it's own module
